# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/problems": {
      "get": {
        "operationId": "ProblemsController_list_v1",
        "parameters": [
          {
            "name": "filter[status]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending_validation",
                "awaiting_sponsorship",
                "funding_required",
                "in_progress",
                "verification_pending",
                "resolved",
                "rejected"
              ]
            }
          },
          {
            "name": "filter[category]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "water",
                "roads",
                "drainage",
                "streetlight",
                "waste",
                "bridge",
                "sanitation",
                "school",
                "health"
              ]
            }
          },
          {
            "name": "filter[departmentId]",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "weights",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List problems in the caller’s jurisdiction, ranked",
        "tags": [
          "problems"
        ]
      }
    },
...
    "/api/v1/problems/{id}/sponsorship/invite": {
      "post": {
        "operationId": "SponsorshipController_invite_v1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Send the same key on a retry and the transition happens once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Invite every matched industry to sponsor this problem",
        "tags": [
          "sponsorship"
        ]
      }
    },
    "/api/v1/problems/{id}/sponsorship/approve": {
      "post": {
        "operationId": "SponsorshipController_approve_v1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Send the same key on a retry and the transition happens once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveSponsorshipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Accept an industry's sponsorship proposal",
        "tags": [
          "sponsorship"
        ]
      }
    },
    "/api/v1/problems/{id}/sponsorship/decline": {
      "post": {
        "description": "When the last sponsor still in play declines, the government funding fallback fires in the same transaction — an officer should not have to notice that industry ran out.",
        "operationId": "SponsorshipController_decline_v1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Send the same key on a retry and the transition happens once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeclineSponsorshipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Record that an industry declined",
        "tags": [
          "sponsorship"
        ]
      }
    },
    "/api/v1/problems/{id}/sponsorship/fallback": {
      "post": {
        "description": "Recommends a source and says whether the department can afford it. It does not approve anything: moving public money stays a decision somebody with funding.approve makes.",
        "operationId": "SponsorshipController_fallback_v1",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Send the same key on a retry and the transition happens once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Abandon sponsorship and fall back to government funding",
        "tags": [
          "sponsorship"
        ]
      }
    },
...
  },
  "info": {
    "title": "Jan Setu API",
    "description": "Generated — do not edit by hand. Run `npm run openapi:gen`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApproveSponsorshipDto": {
        "type": "object",
        "properties": {
          "sponsorId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "amount": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 1000000000000,
            "exclusiveMaximum": false
          }
        },
        "required": [
          "sponsorId"
        ]
      },
      "DeclineSponsorshipDto": {
        "type": "object",
        "properties": {
          "sponsorId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500
          }
        },
        "required": [
          "sponsorId",
          "reason"
        ]
      },
...
    }
  }
}